refactor: rm openapireact + react agents
