--- dapr_agents/agent/patterns/openapi/__init__.py
@@ -1,1 +1,0 @@
-from .react import OpenAPIReActAgent
